Add files via upload
W2V & Glove
PCA calculated on 7,500 + gendered terms.
- Projected neutral words [e.g. human, spouse] onto the above PC1 & PC2
- Projected target words representing pleasant & unpleasant chosen randomly from the WEAT word list [e.g. human, spouse] onto the above PC1 & PC2

Calculated the average Euclidean distance of the neutral term to the pleasant target words & unpleasant target words.

For all steps above, I tried both with NO scaling of the word embeddings & Standard Scaling by feature.

Noticed that there is a clear pattern in the arrow plot representing the male-female-neutral pair for Glove on page 7 left. This pattern is also present in many other plots.

Noticed that both pleasant and unpleasant target words are overall [before CN] significantly closer to male terms as compared to female terms. I would expect there to be an opposite effect w.r.t. male & female terms when looking at pleasant & unpleasant words. This is not the case. However, looking at the target words of both categories, there might be a chance that these emotion/sentiment-strong words are more often associated with male terms as compared to female terms. This needs further testing on some other target word lists.

Also noticed that after CN, there is a global pattern that the L2 distance of the male and female terms to each of the target words shrunk by a lot. What's worth mentioning is that the DIFFERENCE in terms of the average distance of the male terms and female terms to each of the target words shrunk. For standard scaled glove on page 20 right, the result even flipped on the target word "kill". The shrinkage of the difference could be a sign of a reduced gender bias.